requirements.txt
gradio>=4.0.0
torch>=1.9.0
transformers>=4.20.0
datasets>=2.0.0
sentencepiece>=0.1.95
Pillow>=8.0.0
huggingface-hub>=0.10.0
accelerate>=0.20.0
python-dotenv>=0.19.0
